requirements.txt
flask>=2.0.0
insightface>=0.7.3
onnxruntime>=1.14.0
numpy>=1.21.0
pillow>=9.0.0
scipy>=1.7.0